+++ client/assets/js/lib/utils.js
@@ -8,4 +8,14 @@ function addNewHabitToDOM(data) {
 	habits.insertBefore(habit, habits.firstChild);
 }
 
-module.exports = { toggleNav, addNewHabitToDOM };
+function updateHabitDescription() {
+	const form = document.querySelector('form');
+	const description = document.querySelector('.description');
+
+	const name = form.name.value || '*habit*';
+	const goal = form.goal.value || '*goal*';
+
+	description.innerText = `I am going to ${name} ${goal} times per day`;
+}
+
+module.exports = { toggleNav, addNewHabitToDOM, updateHabitDescription };