+++ changelog.md
@@ -7,10 +7,14 @@
   - index
   - dashboard
   - profile
+- Add dynamic habit description to add habit form
 
 ## Backend
+
 [1.0.0] - 2021-07-26 - Poligera & Ridwan Axmed
+
 ### Added
+
 - PostgreSQL database setup with defined tables (users and habits) and seed data
 - Docker-Compose.yaml file
 - Models for User and Habit with crud functionality, complete with db queries
@@ -20,6 +24,5 @@
 - Middleware for jsonwebtoken which verifies validity of the token to authorize user entry
 - .env file with SECRET environment variable used with token verification process
 - package.json file with aproppriate dependencies
-
 
 ## Misc